+++ packages/uilib/src/lib/plugins/network-explorer/diagram-container/diagram-container.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import { extractIEDInfosWithBay, extractIEDNetworkInfoV2 } from "../layout/get-ieds"
+import { extractIEDInfosWithBay, extractIEDNetworkInfoV2, findAllIEDBays } from "../layout/get-ieds"
 import { calculateLayout, calculateLayoutV2, type Config } from "../layout/node-layout"
 import Diagram from "./diagram.svelte"
 import { Position, type Edge, type Node } from "@xyflow/svelte"
@@ -16,9 +16,9 @@ export let root: Element
 // 
 const config: Config = {
 	width: 	200,
-	height: 100,
+	height: 30,
 
-	spacingBase:         100,
+	spacingBase:         10,
 	spacingBetweenNodes: 100,
 }
 
@@ -36,40 +36,73 @@ async function initInfos(
 		return []
 	}
 		
-	const iedInfos = extractIEDInfosWithBay(root)
-	rootNode = await calculateLayout(iedInfos)
+	// const iedInfos = extractIEDInfosWithBay(root)
+	// rootNode = await calculateLayout(iedInfos)
+
 	const iedNetworkInfo = extractIEDNetworkInfoV2(root)
-	rootNode = await calculateLayoutV2(iedNetworkInfo, config)
+	const iedBayMap = findAllIEDBays(root)
+	rootNode = await calculateLayoutV2(iedNetworkInfo, iedBayMap, config)
 
 
-	nodes = rootNode?.children.map(child => {
-		return {
-			...child,
+	for(const node of rootNode.children){
+		// const targetPosition = node.layoutOptions.
+		const newNode = {
+			...node,
 			position: {
-				x: child.x!,
-				y: child.y!,
+				x: node.x!,
+				y: node.y!,
 			},
 			data: {
-				label: child.label,
+				label: node.label,
 			},
-			targetPosition: Position.Right,
-			sourcePosition: Position.Left,
-			// targetPosition: Position.Top,
-			// sourcePosition: Position.Bottom,
+			style: `width: ${node.width}px; height: ${node.height}px;`,
+			// targetPosition: Position.Right,
+			// sourcePosition: Position.Left,
+			targetPosition: Position.Top,
+			sourcePosition: Position.Bottom,
 		}
-	})
+		nodes.push(newNode)
+
+		const hasChildren = node.children?.length ?? 0 > 0
+		if(hasChildren){
+			for(const subNode of node.children){
+				nodes.push({
+					...subNode,
+					parentNode: newNode.id,
+					position: {
+						x: subNode.x!,
+						y: subNode.y!,
+					},
+					data: {
+						label: subNode.label ?? "",
+					},
+					style: `width: ${subNode.width}px; height: ${subNode.height}px;`,
+					targetPosition: Position.Right,
+					sourcePosition: Position.Left,
+					// targetPosition: Position.Top,
+					// sourcePosition: Position.Bottom,
+				})
+			}
+		}
+			
+	}
+
 
 	edges = rootNode?.edges?.map(edge => {
 		return {
 			id:     edge.id,
 			source: edge.sources[0],
 			target: edge.targets[0],
-			type:   "smoothstep",
+			// type:   "smoothstep",
+			// type: "step",
 			// type: "straight",
+			type: "bezier",
+			// animated: true,
+			style: 'stroke: #fff;',
 		}
 	}) ?? []
 
-	console.log({level: "dev", iedInfos, iedNetworkInfo, rootNode, nodes})
+	console.log({level: "dev", iedBayMap, iedNetworkInfo, rootNode, nodes})
 }